Makefile: extract compilation flags to variables
Previously, the WASM compilation step was hard to adjust for different
systems, as it required re-writing the actual rule.

This change extracts all the flags into make variables,
and separates them into system- and project-specific ones.

The system-specific ones are expected to change from machine to machine,
and now can be easily overridden:

    $ make WASM_CC=clang WASM_CFLAGS=--target=wasm32-wasi …

The project-specific flags are also extracted to variables,
although at this point this is mainly for consistency sake.