feat: add initSync() to initialize WASM synchronously
This uses the synchronous WebAssembly APIs to compile the WASM
part synchronously, which allows users to initialize the library
and parse synchronously with WASM without having to resort
to the JS version. This would be useful for Node.js core to
use the WASM version in paths that require synchronous
initialization (it currently always uses the JS version for that).