feat: build wasm in common wasm-builder
Refs: https://github.com/nodejs/security-wg/issues/1236

Update to build in docker using wasm-builder image
maintained in https://github.com/nodejs/wasm-builder.

As a side effect this also updates the versions of the
wasm tools to a newer version.

undici and amaro have already been moved over and I plan to
document in Node.js docs that this is the projects standard
way to build WASM blobs.

Signed-off-by: Michael Dawson <midawson@redhat.com>